Renamed full_test to test_pipeline.py
