+++ Datasnakes/Orthologs/Blast/__init__.py
@@ -1,7 +1,5 @@
-from .comparative_genetics_blastn import CompGenBLASTn
-
+from .blastn import CompGenBLASTn
 from .comparative_genetics_files import CompGenFiles
-
 from .comparative_genetics_objects import CompGenObjects
 
 # Make this explicit, then they show up in the API docs
